Update deploy to Python3

+++ sexy_print.py
@@ -41,10 +41,10 @@ class sexy_print(object):
         text_length = len(text)
         print (80 * "-")
         if (text_length % 2 == 0):
-            padding = (80 - len(text)) / 2 * "-"
+            padding = int((80 - len(text)) / 2) * "-"
             print(padding + self.GREEN + text + self.NORMAL + padding)
         else:
-            padding_left = (80 - len(text) - 1) / 2 * "-"
-            padding_right = (80 - len(text) + 1) / 2 * "-"
+            padding_left = int((80 - len(text) - 1) / 2) * "-"
+            padding_right = int((80 - len(text) + 1) / 2) * "-"
             print(padding_left + self.GREEN + text + self.NORMAL + padding_right)
         print (80 * "-")
